Fix variation id warning for Mixpanel data sources
